refactor: update SectionHeader import to use shared component in blogs and playlist pages; adjust searchParams handling in PlaylistPage
